addressed fix for scary reassignment, now just requires a few hooks
